+++ src/components/AggregateStatusCard.vue
@@ -5,16 +5,11 @@
       'card-pf-aggregate-status-mini': isLayoutMini
     }">
     <h2 class="card-pf-title">
-      <a v-if="href" :href="href">
+      <span :is="href ? 'a' : 'span'" :href="href">
         <span v-if="iconClass" :class="iconClass"></span>
         <span v-if="count" class="card-pf-aggregate-status-count">{{count}}</span>
         <span class="card-pf-aggregate-status-title">{{title}}</span>
-      </a>
-      <template v-else>
-        <span v-if="iconClass" :class="iconClass"></span>
-        <span v-if="count" class="card-pf-aggregate-status-count">{{count}}</span>
-        <span class="card-pf-aggregate-status-title">{{title}} </span>
-      </template>
+      </span>
     </h2>
     <div class="card-pf-body">
       <p class="card-pf-aggregate-status-notifications" v-show="!isLayoutMini || iconClass || count">
@@ -30,7 +25,7 @@ export default {
 
   props: {
     title: String,
-    count: String,
+    count: [String, Number],
     href: String,
     iconClass: String,
     layout: String,